refactor: Move Storybook APIs out from Wrapper
It was so hard to add/modify/reuse the Wrapper component due to its
code highly coupled with Storybook's APIs. This change allos us to
render the Wrapper anywhere, includes in Docs and in stories.
